VietCD commit show wishlist by retrofit 3419

+++ app/src/main/java/com/swp/culturehomestay/models/HomestayImage.java
@@ -17,10 +17,10 @@ public class HomestayImage {
     private String imageUrl;
     @SerializedName("createdDate")
     @Expose
-    private Integer createdDate;
+    private long createdDate;
     @SerializedName("editedDate")
     @Expose
-    private Integer editedDate;
+    private long editedDate;
 
     public String getOrderImage() {
         return orderImage;
@@ -46,19 +46,19 @@ public class HomestayImage {
         this.imageUrl = imageUrl;
     }
 
-    public Integer getCreatedDate() {
+    public long getCreatedDate() {
         return createdDate;
     }
 
-    public void setCreatedDate(Integer createdDate) {
+    public void setCreatedDate(long createdDate) {
         this.createdDate = createdDate;
     }
 
-    public Integer getEditedDate() {
+    public long getEditedDate() {
         return editedDate;
     }
 
-    public void setEditedDate(Integer editedDate) {
+    public void setEditedDate(long editedDate) {
         this.editedDate = editedDate;
     }
